Updated config files
