+++ auto_workflow/config.py
@@ -19,7 +19,6 @@ OUTPUT_DIR = 'output'
 # tdb should set `tdb:unionDefaultGraph true`
 FUSEKI_URL = 'http://localhost:3030'
 GRAPH_BASE_URL = 'http://data.americanartcollaborative.org/'
-# GRAPH_BASE_URL = 'http://data.aac.org/'
 # FUSEKI_ACCOUNT = ('', '')
 FUSEKI_ACCOUNT = None
 TDB_DATASET_NAME = 'american-art'
